kernel - Added missing includes, fixed some debug output while debugging, define checks and added more debug output - Disabled neon register backup and cache handling as both are not yet ready to work - Replaced magic numbers in push / pop fpu register by define usage - Fixed set of spsr within thread creation - Adjusted interrupt handler macro stack adjustment and secure mode return - Fixed task_thread_switch_to to work correctly - Transformed uint32_t to uintptr_t within mask interrupt helper - Reduced amount of interrupts per second to one while testing on real hardware - Added option to print origin of stack smashing issue - Added empty assembly files for adding optimized memcyp, memset and strlen - Removed unused strcmp implementation from kernel - Revised extend_heap_space to await given_size as minimum necessary for extension to prevent recursion hell - Commented out enable_release handling for kernel as it needs to be checked again
tool
- Started working on simple script to mount sd card and copy files to correct place

server
- Replaced usage of fork by _process_fork for vfs within boot
- Added optional sleep wait to server helper send vfs add request
- Extended servers generally by more print during start
- Changed framebuffer driver to force width and height from define
- Temporarily added -Wl,-s and -Wl,--gc-sections to vfs

+++ bolthur/kernel/platform/rpi/interrupt.c
@@ -25,6 +25,7 @@
 #include <panic.h>
 #include <platform/rpi/gpio.h>
 #include <platform/rpi/peripheral.h>
+#include <debug/debug.h>
 
 /**
  * @brief Helper to validate interrupt number
@@ -55,10 +56,10 @@ bool interrupt_validate_number( size_t num ) {
 void interrupt_mask_specific( int8_t num ) {
   uint32_t interrupt = ( uint32_t )num;
   // get peripheral base
-  uint32_t base = ( uint32_t )peripheral_base_get( PERIPHERAL_GPIO );
+  uintptr_t base = peripheral_base_get( PERIPHERAL_GPIO );
   // get interrupt enable and pending
-  uint32_t interrupt_enable = base;
-  uint32_t interrupt_pending = base;
+  uintptr_t interrupt_enable = base;
+  uintptr_t interrupt_pending = base;
   if ( 32 > interrupt ) {
     interrupt_enable += INTERRUPT_ENABLE_IRQ_1;
     interrupt_pending += INTERRUPT_IRQ_PENDING_1;
